+++ include/ev/misc.h
@@ -115,6 +115,12 @@ void ev_buf_make_v(ev_buf_t bufs[], size_t nbuf, va_list ap);
 void ev_library_shutdown(void);
 
 /**
+ * @brief Returns the current high-resolution real time in microsecond.
+ * @return Time in microsecond.
+ */
+uint64_t ev_hrtime(void);
+
+/**
  * @} EV_MISC
  */